[#239] Split test_partijen.py in several tests files
