+++ src/openklant/components/klantinteracties/openapi.yaml
@@ -2433,7 +2433,13 @@ paths:
           description: ''
     post:
       operationId: partijenCreate
-      description: Maak een partij aan.
+      description: "\n**Warnings:**\n\nHandles `partijIdentificatoren` creation with\
+        \ atomicity guarantees.\n\n- If the `UUID` is provided in the `PartijIdentificator`\
+        \ object,\nthe endpoint will treat it as an update operation for the existing\
+        \ `PartijIdentificator`,\napplying the provided data and linking the parent\
+        \ `Partij` to the new one created.\n- If the `UUID` is **not** specified,\
+        \ the system will create a new\n`PartijIdentificator` instance respecting\
+        \ all uniqueness constraints.\n    "
       summary: Maak een partij aan.
       tags:
       - partijen
@@ -2493,7 +2499,12 @@ paths:
           description: ''
     put:
       operationId: partijenUpdate
-      description: Werk een partij in zijn geheel bij.
+      description: "\n**Warnings:**\n\nHandles `partijIdentificatoren` updates with\
+        \ atomicity guarantees.\n\n- If the `UUID` is provided in the `PartijIdentificator`\
+        \ object,\nthe system will update the specified instance with the new data.\n\
+        - If the `UUID` is **not** specified, the system will `DELETE` all `PartijIdentificator`\n\
+        objects related to the parent and `CREATE` new ones with the new passed data.\n\
+        \    "
       summary: Werk een partij in zijn geheel bij.
       parameters:
       - in: path
@@ -2521,7 +2532,12 @@ paths:
           description: ''
     patch:
       operationId: partijenPartialUpdate
-      description: Werk een partij deels bij.
+      description: "\n**Warnings:**\n\nHandles `partijIdentificatoren` updates with\
+        \ atomicity guarantees.\n\n- If the `UUID` is provided in the `PartijIdentificator`\
+        \ object,\nthe system will update the specified instance with the new data.\n\
+        - If the `UUID` is **not** specified, the system will `DELETE` all `PartijIdentificator`\n\
+        objects related to the parent and `CREATE` new ones with the new passed data.\n\
+        \    "
       summary: Werk een partij deels bij.
       parameters:
       - in: path
@@ -5782,6 +5798,24 @@ tags:
 - name: categorieën
 - name: digitale adressen
 - name: interne taken
+  description: |
+    Deze API publiceert notificaties op het kanaal `internetaken`.
+
+    **Main resource**
+
+    `internetaak`
+
+
+
+    **Kenmerken**
+
+    * `nummer`: Uniek identificerend nummer dat tijdens communicatie tussen mensen kan worden gebruikt om de specifieke interne taak aan te duiden.
+    * `gevraagde_handeling`: Handeling die moet worden uitgevoerd om de taak af te ronden.
+    * `toelichting`: Toelichting die, aanvullend bij de inhoud van het klantcontact dat aanleiding gaf tot de taak en de gevraagde handeling, bijdraagt aan het kunnen afhandelen van de taak.
+    * `status`: Aanduiding van de vordering bij afhandeling van de interne taak.
+
+    **Resources en acties**
+    - `internetaak`: create, update, destroy
 - name: klanten contacten
 - name: onderwerpobjecten
 - name: partij-identificatoren
@@ -5802,7 +5836,7 @@ tags:
     * `soort_partij`: Geeft aan van welke specifieke soort partij sprake is.
 
     **Resources en acties**
-    - `partij`: create, update, destroy
+- name: partijen
   description: |2
 
     **Atomicity in Partij and PartijIdentificator**
@@ -5814,21 +5848,20 @@ tags:
 
     For `POST`, `PATCH`, and `PUT` requests for `Partij`,
     it is possible to send a list of `PartijIdentificator` objects.
+    Deze API publiceert notificaties op het kanaal `partijen`.
 
-    **Warnings:**
+    **Main resource**
 
-    - In all requests, `PartijIdentificator` objects should not contain the **UUID**
-    of the parent `Partij`, because it is automatically assigned.
-    - `POST` request:
-        - If the **UUID** is provided in the `PartijIdentificator` object,
-        the endpoint will treat it as an update operation for the existing `PartijIdentificator`,
-        applying the provided data and linking the parent `Partij` to the new one created.
-        - If the **UUID** is **not** specified, the system will create a new
-        `PartijIdentificator` instance respecting all uniqueness constraints.
-    - `PATCH` or `PUT` requests:
-        - If the **UUID** is provided in the `PartijIdentificator` object,
-        the system will update the specified instance with the new data.
-        - If the **UUID** is **not** specified, the system will `DELETE` all `PartijIdentificator`
-        objects related to the parent and `CREATE` new ones with the passed data.
+    `partij`
+
+
+
+    **Kenmerken**
+
+    * `nummer`: Uniek identificerend nummer dat tijdens communicatie tussen mensen kan worden gebruikt om de specifieke partij aan te duiden.
+    * `interne_notitie`: Mededelingen, aantekeningen of bijzonderheden over de partij, bedoeld voor intern gebruik.
+    * `soort_partij`: Geeft aan van welke specifieke soort partij sprake is.
+
+    **Resources en acties**
 - name: rekeningnummers
 - name: vertegenwoordigingen
